+++ www/script/display_utils.py
@@ -1,0 +1,5 @@
+from pyscript import display
+from utils import display_a_file
+
+display_a_file("./utils.py", "code-utils")
+display("Utils files", target="utils-title")